Tidy up + DatabaseUtilities tests

+++ core/src/main/java/com/tealium/core/internal/persistence/DatabaseUtilities.kt
@@ -35,6 +35,8 @@ internal fun SQLiteDatabase.transaction(
  * Drops the table with the provided [tableName].
  */
 internal fun SQLiteDatabase.dropTable(tableName: String) {
+    if (tableName.isBlank()) return
+
     execSQL(
         "DROP TABLE $tableName"
     )
@@ -44,6 +46,8 @@ internal fun SQLiteDatabase.dropTable(tableName: String) {
  * Safely drops the table with the provided [tableName] if it exists.
  */
 internal fun SQLiteDatabase.dropTableIfExists(tableName: String) {
+    if (tableName.isBlank()) return
+
     execSQL(
         "DROP TABLE IF EXISTS $tableName"
     )
